requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
Pillow==6.2.1
pyasn1==0.4.7
pycryptodome==3.9.2
PyExecJS==1.5.1
requests==2.22.0
rsa==4.0
six==1.13.0
tornado==6.0.3
urllib3==1.25.6
